HIP04: Load bootwrapper when exiting boot services
Today we load the bootwrapper binaries before we enter a kernel using
direct kernel boot. But kernels loaded via EFI applications (like grub2)
also need to have the bootwrapper code loaded.

This patch adds a callback when we exit the runtime services and in that
case loads the bootwrapper code. This happens with direct kernel boot as
well as with grub2 boot.

Signed-off-by: Alexander Graf <agraf@suse.de>