HIP04: Run bootwrapper before we exit runtime services
When we exit the runtime services, we need to also execute our bootwrapper
code to make sure that we get into HYP mode.

This patch adds wrapping code that allows us to execute the bootwrapper code
when exiting runtime services. However we only run them for non-direct kernel
boot, as that one would run the bootwrapper twice otherwise.

Signed-off-by: Alexander Graf <agraf@suse.de>