HIP04: Change architected timer frequency to 50Mhz
When running with HYP mode enabled, Linux uses the architected timer. It detects
168.75Mhz as indicated by firmware, but the timer only runs at 50Mhz. I've tried
to change the speed setting in bootwrapper but the timer was still stuck at 50Mhz.

This patch resignates to hardware and simply configured the architected timer to
be at 50Mhz which makes my Linux guest happily run "sleep 1" within one second.

Signed-off-by: Alexander Graf <agraf@suse.de>